--- a/src/node/audio_buffer_source.rs
+++ b/src/node/audio_buffer_source.rs
@@ -543,7 +543,7 @@ impl AudioProcessor for AudioBufferSourceRenderer {
 
                                 (1. - k) * prev_sample + k * next_sample
                             }
-                            None => 0.
+                            None => 0.,
                         };
                     });
             });