… examples/benchmarks.rs