Hey,

A first progress in improving `AudioScheduledSourceNode` performances

Mostly reviewed how output is created:
- fill `output` channel by channel to access memory in an aligned way
- uses more functional paradigms

This improves perf between 1.3 and 1.5 in benchmarks on my mac, which I think is really outside the noise

### before

| id      | name                                                                     | duration (ms) | Speedup vs. realtime  | buffer.duration (s) |
| - | - | - | - | - |
| 2       | Simple source test without resampling (Mono)                             | 147           | 816.3x                | 120 |
| 3       | Simple source test without resampling (Stereo)                           | 193           | 621.8x                | 120 |
| 4       | Simple source test with resampling (Mono)                                | 145           | 827.6x                | 120 |
| 5       | Simple source test with resampling (Stereo)                              | 195           | 615.4x                | 120 |

### after

| id      | name                                                                     | duration (ms) | Speedup vs. realtime  | buffer.duration (s)
| - | - | - | - | - |
| 2       | Simple source test without resampling (Mono)                             | 98            | 1224.5x               | 120
| 3       | Simple source test without resampling (Stereo)                           | 141           | 851.1x                | 120
| 4       | Simple source test with resampling (Mono)                                | 98            | 1224.5x               | 120
| 5       | Simple source test with resampling (Stereo)                              | 140           | 857.1x                | 120

I would be interested if you see the same kind of changes on your side ?

In any case, do not merge yet, I would also like to try something else.

I think we could also apply that strategy to the delay node (which also would allow to remove `set_channels_values_at` fomr `AudioRenderQuantum`)

### edit after 023e767

| id      | name                                                                     | duration (ms) | Speedup vs. realtime  | buffer.duration (s)
| - | - | - | - | - |
| 2       | Simple source test without resampling (Mono)                             | 90            | 1333.3x               | 120
| 3       | Simple source test without resampling (Stereo)                           | 108           | 1111.1x               | 120
| 4       | Simple source test with resampling (Mono)                                | 91            | 1318.7x               | 120
| 5       | Simple source test with resampling (Stereo)                              | 110           | 1090.9x               | 120