@orottier 2022-07-04
Hey nice, I'm seeing improvements too

```
- 2       | Simple source test without resampling (Mono)                             | 59            | 2033.9x               | 120
- 3       | Simple source test without resampling (Stereo)                           | 67            | 1791.0x               | 120
- 4       | Simple source test without resampling (Stereo and positionnal)           | 196           | 612.2x                | 120
- 5       | Simple source test with resampling (Mono)                                | 54            | 2222.2x               | 120
- 6       | Simple source test with resampling (Stereo)                              | 66            | 1818.2x               | 120
- 7       | Simple source test with resampling (Stereo and positionnal)              | 191           | 628.3x                | 120
```

Compared to current main
```
- 2       | Simple source test without resampling (Mono)                             | 61            | 1967.2x               | 120
- 3       | Simple source test without resampling (Stereo)                           | 82            | 1463.4x               | 120
- 4       | Simple source test without resampling (Stereo and positionnal)           | 203           | 591.1x                | 120
- 5       | Simple source test with resampling (Mono)                                | 57            | 2105.3x               | 120
- 6       | Simple source test with resampling (Stereo)                              | 83            | 1445.8x               | 120
- 7       | Simple source test with resampling (Stereo and positionnal)              | 209           | 574.2x                | 120
```

Exciting news - moreover when we can apply it to other nodes.

I have not looked into the code changes yet, will do tomorrow